UI: Replaced acceptance tests with feature tests
