+++ frontend/tests/acceptance/steps/sign-in-steps.js
@@ -1,0 +1,17 @@
+import steps from './steps';
+import ENV from 'demo/config/environment';
+import { default as mockedWindow } from 'ember-window-mock';
+
+// step definitions that are shared between features should be moved to the
+// tests/acceptance/steps/steps.js file
+
+export default function(assert) {
+  return steps(assert)
+    .then('I should be redirected to the "sign-in" page', function() {
+      assert.stringStartsWith(
+        mockedWindow.location.href,
+        `${ENV.APP.demoOAuthUrl}/authorize`,
+        this.step
+      );
+    });
+}
